requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-experimental
langchain-openai
openai
tabulate
